commit without folder "resources"

+++ src/main/webapp/root/js/controllers/buttons.js
@@ -1,44 +1,30 @@
-'use strict';
-angular.module('buttons', ['ui.bootstrap']);
-var ButtonsController= function ($scope) {
-    	/**********************************************************************************************************************************************/
-    	// These functions regards the grid row selection
-    								$scope.backButton = function(size) {
-    									$location.path("/homePage");
-    								};
-    								
-    				$scope.selectRowButton = function(size) {
-    								type="modify";
-    								if ($scope.mySelections[0] == null
-    										|| $scope.mySelections[0] == "") {
-    									alert("No activity has been selected!");
-    								} else {				
-    									var ben = $scope.mySelections[0].activity.personActivities[1];
-    									retrievePeople($scope.mySelections[0].activity.personActivities[0],size);
-    									if(ben!=null && ben!='')retrievePeople(ben,size);
-    													
-    								}
-    								//$scope.activity = $scope.mySelections[0].activity;
-    							};
+angular.module('buttons', [ 'ui.bootstrap' ]).controller("ButtonsController"),
+		[
+				'$scope','$location',
+				function($scope, $location) {
+					/** ******************************************************************************************************************************************* */
+					// These functions regards the grid row selection
+					$scope.backButton = function(size) {
+						$location.path("/homePage");
+					};
 
-    							$scope.deleteRowButton = function(size) {
-    								type="delete";
-    								if ($scope.mySelections[0] == null
-    										|| $scope.mySelections[0] == "") {
-    									alert("No person has been selected!");
-    								} else {
-    									type="delete";
-    									//$scope.activity = $scope.mySelections[0].activity;
-    									openModal(size);					
-    								}
-    							};
+					$scope.deleteRowButton = function(size) {
+						type = "delete";
+						if ($scope.mySelections[0] == null
+								|| $scope.mySelections[0] == "") {
+							alert("No row has been selected!");
+						} else {
+							type = "delete";
+							openModal(size);
+						}
+					};
 
-    							// ROW INSERT
-    							$scope.insertRowButton = function(size) {
-    								type="insert";
-    								openModal(size);			 				
-    							};
-    							
-    	/**********************************************************************************************************************************************/		
+					// ROW INSERT
+					$scope.insertRowButton = function(size) {
+						type = "insert";
+						openModal(size);
+					};
 
-    };
+					/** ******************************************************************************************************************************************* */
+
+				} ];
